pyproject.toml
[project]
name = "stremlit-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.80.0",
    "openai-agents>=0.6.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
